% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_hugo}
\alias{build_hugo}
\title{Build a Hugo site}
\usage{
build_hugo(with_rmd = TRUE, rmd_folder = "R/Rmd", rebuild = FALSE)
}
\arguments{
\item{with_rmd}{logical flag for whether to build Rmd files}

\item{rmd_folder}{path to the folder containing Rmd files}
}
\description{
\code{build_hugo()} Hugo-ifies Rmd files and then calls Hugo to build the
static site.
}
