% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{clear_filters}
\alias{clear_filters}
\title{Clear crosstalk filters}
\usage{
clear_filters(text = "Clear filters")
}
\arguments{
\item{text}{The text to display (defaults to "Clear filters")}
}
\description{
Insert a "clear filters" option to a set of filter controls. This adds an
HTML \code{div} element that works with jQuery in govukhugo to clear crosstalk
filters on a page, note that this clears all filters on a page including
any DT search boxes.
}
