% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dir.R
\name{data_dir}
\alias{data_dir}
\title{Get the right data directory}
\usage{
data_dir(alt_path = NULL)
}
\arguments{
\item{alt_path}{a non-standard location}
}
\value{
the location of the data folder
}
\description{
The govukhugo package assumes data is stored in R/data, when processing
Rmarkdown files this folder is copied to a temporary folder. This is a helper
function to easily work both interactively when editing an Rmarkdown file
and when running the \code{build_hugo_rmd()} function.
}
