% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govuk_datatable.R
\name{govuk_datatable}
\alias{govuk_datatable}
\title{datatables for GOV.UK}
\usage{
govuk_datatable(
  data,
  title = NULL,
  element_id = NULL,
  col_names = NULL,
  page_length = 10,
  search = FALSE,
  small_text = FALSE,
  buttons = TRUE,
  col_defs = NULL,
  copy_info = NULL,
  export_file = NULL,
  options = NULL
)
}
\arguments{
\item{data}{the data frame to display}

\item{title}{a title/caption for the table}

\item{element_id}{optionally, an id for the resulting HTML chunk}

\item{col_names}{optionally, a vector of column names, otherwise the names of the data object}

\item{page_length}{the number of rows to show per page, default is 10}

\item{search}{whether to include the search box}

\item{small_text}{whether to render the table content with a smaller font}

\item{buttons}{whether to include copy and download buttons}

\item{col_defs}{optionally, a list of column definitions}

\item{copy_info}{information to include when copying the file}

\item{export_file}{the file name for the exported csv}

\item{options}{optionally, a list of additional options to pass to \code{\link[DT:datatable]{DT::datatable()}}}
}
\description{
\code{govuk_datatable} is a wrapper for \code{\link[DT:datatable]{DT::datatable()}} that will enable the
govuk-hugo theme to apply GOV.UK CSS classes to the output.
}
