% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{govuk_palette}
\alias{govuk_palette}
\title{Use GOV.UK colours}
\usage{
govuk_palette(pal = "categorical")
}
\arguments{
\item{pal}{One of \code{categorical} (the default), \code{blue}, \code{blrd}, \code{bldrd_dark},
\code{blyl}, \code{putq}, or a colour name from \code{govuk_colours}}
}
\value{
A set of hexadecimal colours
}
\description{
Access the GOV.UK design system colour palette
}
\details{
The GOV.UK design system includes a colour palette, these can be accessed
via the \code{govuk_colours} vector, it includes both Sass variable colour names
(e.g. \code{govuk-text-colour}) as well as the named colour palette.

The GOV.UK colour scheme was not designed with data visualisation in mind.
The \code{govuk_palette} function provides an opinionated selection of GOV.UK
colours for use in charts (e.g. via \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}}).

The \code{categorical} palette (the default) provides an opinonated set of six
GOV.UK colours for use in categorical/qualitative palettes.

The \code{blues} palette orders the GOV.UK colour palettes blues from dark to
light and can be used for sequential colour palettes. Alternatively you can
supply a name from the \code{govuk_colours} vector to generate the end-points for
a sequential palette using that colour (where the lightest colour is roughly
a third lighter than the input colour).

There are also four opinionated palettes for use in a divergent scale, these
all use an off-white (the mid-colour between GOV.UK light grey and white)
\itemize{
\item \code{blrd} which use GOV.UK blue and GOV.UK red as its end-points
\item \code{blrd_dark} which uses GOV.UK dark blue as its blue end-point and a darker
red of the same hue as GOV.UK red as its red end-point
\item \code{blyl} which uses GOV.UK blue and GOV.UK yellow as its end-points
\item \code{putq} which uses GOV.UK purple and GOV.UK turquoise as its end-points
}

The categorical and divergent palettes have been checked for general support
for users with colour blindness.
}
