% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_govuk_hugo}
\alias{init_govuk_hugo}
\title{Initiate a Hugo site using the govuk-hugo theme}
\usage{
init_govuk_hugo()
}
\description{
\code{init_govuk_hugo()} calls Hugo to create a new site scaffold,
and adds govuk-hugo as a theme. Note that this function only
works inside of RStudio with an RStudio project. It is
recommended that your project also uses git for version control.
}
