% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_rmd}
\alias{render_rmd}
\title{Hugo-ify an R Markdown document}
\usage{
render_rmd(rmd_file, tmp_dir = tempdir(), out_dir = NULL)
}
\arguments{
\item{rmd_file}{the .rmd file to convert}

\item{tmp_dir}{a temporary directory for processing}

\item{out_dir}{override the default output directory}
}
\description{
Hugo is largely used to convert plain markdown files into HTML, but it also
supports HTML documents as input files. \code{render_rmd()} renders Rmd files
using \code{\link[rmarkdown:html_document]{rmarkdown::html_document()}} and the converts the HTML to work with
Hugo.

By default \code{render_rmd()} will read the YAML front-matter of an Rmd for
a section parameter and use that as a sub-directory of the content folder
as the output location.
}
