% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.R
\name{serve_site}
\alias{serve_site}
\alias{stop_server}
\title{Serve the site}
\usage{
serve_site(
  build_rmd = TRUE,
  rmd_folder = "R/Rmd",
  rebuild = FALSE,
  drafts = TRUE,
  future = TRUE
)

stop_server()
}
\arguments{
\item{build_rmd}{whether to build Rmd files before serving}

\item{rmd_folder}{the folder containing Rmd files}

\item{rebuild}{whether to trigger a rebuild of all Rmd files}

\item{drafts}{whether to include drafts}

\item{future}{whether to include pages with a date in the future}
}
\description{
These are wrappers around \code{\link[blogdown:serve_site]{blogdown::serve_site()}} and \code{\link[blogdown:serve_site]{blogdown::stop_server()}}.
The main purpose is to ensure that the blogdown re-knitting process doesn't
run when files are saved. \code{stop_server()} is a simple re-export.
}
\section{Functions}{
\itemize{
\item \code{stop_server}: Stop serving a Hugo site
}}

