% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_hugo_rmd}
\alias{build_hugo_rmd}
\title{Bulk Hugo-ify Rmd files}
\usage{
build_hugo_rmd(rmd_folder = "R/Rmd", rebuild = FALSE)
}
\arguments{
\item{rmd_folder}{path to the folder containing Rmd files}
}
\description{
\code{build_hugo_rmd()} converts all Rmd files within a folder to Hugo-ified
HTML, it will place the Hugo-ified HTML files into the content sections
designated in the Rmd YAML front matter. Use \code{\link[=render_rmd]{render_rmd()}} if you want
to override this. By default it expects Rmd files in the \code{R/Rmd} folder.
}
