% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{new_rmd}
\alias{new_rmd}
\title{Open a new editor with the Rmarkdown skeleton}
\usage{
new_rmd()
}
\description{
The govukhugo package includes a skeleton Rmarkdown file,
which can be easily selected via the RStudio GUI. \code{new_rmd()}
allows you to open a new RStudio editor with a copy of the
skeleton via the console.
}
