% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.R
\name{serve_site}
\alias{serve_site}
\title{Serve the site}
\usage{
serve_site(
  ...,
  .site_dir = NULL,
  build_rmd = TRUE,
  rmd_folder = "R/Rmd",
  rebuild = FALSE
)
}
\arguments{
\item{...}{Arguments passed to \code{\link[servr:server_config]{servr::server_config()}}}

\item{.site_dir}{Directory to search for site configuration file, defaults to \code{\link[base:getwd]{base::getwd()}}.}
}
\description{
This is a wrapper around \code{\link[blogdown:serve_site]{blogdown::serve_site()}}. It's main
purpose is to ensure that the blogdown re-knitting process doesn't
run when files are saved.
}
