% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.R
\name{serve_site}
\alias{serve_site}
\alias{stop_server}
\title{Serve the site}
\usage{
serve_site(
  build_rmd = TRUE,
  rmd_folder = "R/Rmd",
  rebuild = FALSE,
  drafts = TRUE,
  future = TRUE
)

stop_server()
}
\arguments{
\item{...}{Arguments passed to \code{\link[servr:server_config]{servr::server_config()}}}

\item{.site_dir}{Directory to search for site configuration file, defaults to \code{\link[base:getwd]{base::getwd()}}.}
}
\description{
These are wrappers around \code{\link[blogdown:serve_site]{blogdown::serve_site()}} and \code{\link[blogdown:serve_site]{blogdown::stop_server()}}.
The main purpose is to ensure that the blogdown re-knitting process doesn't
run when files are saved. \code{stop_server()} is a simple re-export.
}
\section{Functions}{
\itemize{
\item \code{stop_server}: Stop serving a Hugo site
}}

